Convert RSS and sitemap generation to build-time static files
This change improves the site's performance on Vercel by:
1. Creating static RSS feed and sitemap XML files during build time
2. Redirecting the dynamic API routes to the static files
3. Reducing resource usage on each request
