+++ package.json
@@ -9,7 +9,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+    "build": "node scripts/generate-static-files.js && next build",
     "start": "next start"
   },
   "dependencies": {
@@ -17,11 +17,12 @@
     "@heroicons/react": "^1.0.1",
     "@mailchimp/mailchimp_marketing": "^3.0.48",
     "@tailwindcss/jit": "^0.1.18",
-    "@tailwindcss/typography": "^0.5.13",
+    "@tailwindcss/typography": "^0.5.15",
     "@tryghost/content-api": "^1.5.8",
     "autoprefixer": "^10.2.6",
     "axios": "^0.21.1",
     "better-react-mathjax": "^1.0.2",
+    "dotenv": "^16.0.3",
     "eslint-config-next": "^14.2.4",
     "fathom-client": "^3.2.0",
     "fs-extra": "^10.0.0",
