+++ pages/rss.js
@@ -1,97 +1,14 @@
-import { getPosts } from './api/ghost_data'
-
-const options = {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-    weekday: 'short',
-    hour: 'numeric',
-    minute: 'numeric',
-    second: 'numeric',
-    timeZone: 'GMT',
-}
-
-const createRSS = (posts) => `<?xml version="1.0" encoding="UTF-8"?>
-    <rss 
-        xmlns:dc="http://purl.org/dc/elements/1.1/" 
-        xmlns:content="http://purl.org/rss/1.0/modules/content/" 
-        xmlns:atom="http://www.w3.org/2005/Atom" 
-        version="2.0" 
-        xmlns:media="http://search.yahoo.com/mrss/"
-    >
-        <channel>
-            <title><![CDATA[ Who is Nnamdi? ]]></title>
-            <description>
-                <![CDATA[ Thoughts on technology, venture capital, and the economics of both ]]>
-            </description>
-            <link>https://whoisnnamdi.com</link>
-            <lastBuildDate>${new Intl.DateTimeFormat('en-GB', options).format(new Date()) + " GMT"}</lastBuildDate>
-            <atom:link href="https://whoisnnamdi.com/rss/" rel="self" type="application/rss+xml"/>
-            ${posts
-                .map((post) => {
-                    return `<item>
-                        <title>
-                            <![CDATA[${post.title}]]>
-                        </title>
-                        <description>
-                            <![CDATA[${post.excerpt}]]>
-                        </description>
-                        <link>${`https://whoisnnamdi.com/${post.slug}/`}</link>
-                        <guid isPermaLink="false">${post.id}</guid>
-                        ${post.tags ?
-                            post
-                            .tags.map((tag) => {
-                                return `<category>
-                                        <![CDATA[ ${tag.name} ]]>
-                                    </category>
-                                `
-                            })
-                            .join("")
-                            :
-                            null
-                        }
-                        <dc:creator>
-                            <![CDATA[ Nnamdi Iregbulem ]]>
-                        </dc:creator>
-                        <pubDate>${new Intl.DateTimeFormat('en-GB', options).format(new Date(post.published_at)) + " GMT"}</pubDate>
-                        <media:content url="${post.feature_image}" medium="image" />
-                        <content:encoded>
-                            <![CDATA[ ${post.html} ]]>
-                        </content:encoded>
-                    </item>
-                    `
-                })
-                .join("")
-            }
-        </channel>
-    </rss>
-`
-
-// This function just returns a empty component
+// This page redirects to the static RSS feed
 export default function RSS() {
     return null
 }
 
-// Use getServerSideProps to handle the request
 export async function getServerSideProps({ res }) {
-    // Set the content type to XML
-    res.setHeader('Content-Type', 'text/xml')
+    // Redirect to the static RSS feed
+    res.statusCode = 301
+    res.setHeader('Location', '/rss/feed.xml')
+    res.end()
     
-    try {
-        // Generate the RSS feed
-        const posts = await getPosts()
-        const rssContent = createRSS(posts)
-        
-        // Write the response directly
-        res.write(rssContent)
-        res.end()
-    } catch (error) {
-        console.error('Error generating RSS feed', error)
-        res.statusCode = 500
-        res.end('Error generating RSS feed')
-    }
-    
-    // Return empty props (this won't be used since we've already ended the response)
     return {
         props: {},
     }
