+++ scripts/test-rss.js
@@ -1,0 +1,6 @@
+// This script tests the RSS generation
+require('dotenv').config();
+const generateRSS = require('./generate-rss');
+
+console.log('Testing RSS generation...');
+generateRSS();
